+++ HealthDisplayWithFont/HealthDisplayWithFontClass.cs
@@ -4,15 +4,13 @@ using Il2CppRUMBLE.Players;
 using Il2CppRUMBLE.Players.Subsystems;
 using Il2CppTMPro;
 using MelonLoader;
-using RumbleModdingAPI;
-using System.Collections;
 using UnityEngine;
 using UnityEngine.TextCore.LowLevel;
 
 // Test with clones
-// The issue only occurs after the playerhealth is initlizied
+// TODO: make gh action to make sure version nums are correct
 
-[assembly: MelonInfo(typeof(HealthDisplayWithFontClass), "HealthDisplayWithFont", "0.1.2", "ninjaguardian", "https://thunderstore.io/c/rumble/p/ninjaguardian/HealthDisplayWithFont")]
+[assembly: MelonInfo(typeof(HealthDisplayWithFontClass), "HealthDisplayWithFont", "0.1.3", "ninjaguardian", "https://thunderstore.io/c/rumble/p/ninjaguardian/HealthDisplayWithFont")]
 [assembly: MelonGame("Buckethead Entertainment", "RUMBLE")]
 
 [assembly: MelonColor(255, 0, 160, 230)]
@@ -30,7 +28,6 @@ namespace HealthDisplayWithFont
     public class HealthDisplayWithFontClass : MelonMod
     {
         private static TMP_FontAsset fontAsset;
-        private static TextMeshPro localHealthBarText;
 
         /// <inheritdoc/>
         public override void OnLateInitializeMelon()
@@ -70,101 +67,84 @@ namespace HealthDisplayWithFont
             {
                 MelonLogger.Warning($"Error loading font bundle. Using default font. Error: {ex.Message}");
             }
-
-            Calls.onMapInitialized += MapInit;
         }
 
-        private static void MapInit()
+        private static void AddHealthbarText(Transform UIBAR, Player player)
         {
-            MelonCoroutines.Start(LocalReloadHealthBar());
-        }
-
-        private static IEnumerator LocalReloadHealthBar()
-        {
-            const byte maxTimeout = 10; //max 255 seconds
-            byte timeout = maxTimeout;
-            while (Calls.Players.GetPlayerController().transform.Find("UI/LocalUI/Local UI Bar") == null || localHealthBarText != null) 
-            {
-                yield return new WaitForSeconds(1);
-                if (--timeout == 0) {
-                    MelonLogger.Error($"Timeout while waiting for local health bar game object. Waited {maxTimeout} seconds.");
-                    MelonLogger.Warning($"Is PlayerController null: {Calls.Players.GetPlayerController() == null} (expected: false)");
-                    MelonLogger.Warning($"Is UI Bar null: {Calls.Players.GetPlayerController().transform.Find("UI/LocalUI/Local UI Bar") == null} (expected: false)");
-                    MelonLogger.Warning($"Is Local UI Bar Text null: {localHealthBarText == null} (expected: true)");
-                    yield break;
-                }
-            }
-
-            GameObject healthBar = new("TextMeshPro");
-            healthBar.transform.SetParent(Calls.Players.GetPlayerController().transform.Find("UI/LocalUI/Local UI Bar"), false);
+            GameObject healthBar = new("HealthText");
+            healthBar.transform.SetParent(UIBAR, false);
 
             TextMeshPro textRef = healthBar.AddComponent<TextMeshPro>();
+
             healthBar.transform.localPosition = new Vector3(-1.01f, 0.01f, 0.1f);
             healthBar.transform.localRotation = Quaternion.Euler(63f, 270f, 0f);
             healthBar.transform.localScale = new Vector3(0.015f, 0.007f, 0.015f);
-            textRef.text = Calls.Players.GetLocalPlayer().Data.HealthPoints.ToString();
+
+            textRef.text = player.Data.HealthPoints.ToString();
 
             if (fontAsset != null)
             {
                 textRef.font = fontAsset;
             }
-
-            localHealthBarText = textRef;
         }
 
-        //[HarmonyPatch(typeof(PlayerHealth), nameof(PlayerHealth.Initialize))]
-        //class PlayerHealthInitPatch
-        //{
-        //    static void Postfix(ref PlayerHealth __instance, ref PlayerController controller)
-        //    {
-        //        MelonLogger.Warning(controller.assignedPlayer.Data.GeneralData.PublicUsername);
+        private static Transform GetHealthbar(Transform UI, ControllerType? controllerType)
+        {
+            if (controllerType == null) return null;
+            if (controllerType == ControllerType.Local)
+            {
+                Transform healthbar = UI.GetChild(0)?.GetChild(1);
+                if (healthbar?.name != "Local UI Bar" || healthbar?.gameObject?.active != true)
+                {
+                    MelonLogger.Warning("Could not get Local Healthbar via GetChild");
+                    healthbar = UI.Find("LocalUI/Local UI Bar");
+                    if (healthbar == null || healthbar.gameObject?.active != true)
+                    {
+                        MelonLogger.Error("Could not get Local Healthbar via Find");
+                        return null;
+                    }
+                }
+                return healthbar;
+            }
+            else if (controllerType == ControllerType.Remote)
+            {
+                Transform healthbar = UI.GetChild(1)?.GetChild(0);
+                if (healthbar?.name != "Remote UI Bar" || healthbar?.gameObject?.active != true)
+                {
+                    MelonLogger.Warning("Could not get Remote Healthbar via GetChild");
+                    healthbar = UI.Find("RemoteUI/Remote UI Bar");
+                    if (healthbar == null || healthbar.gameObject?.active != true)
+                    {
+                        MelonLogger.Error("Could not get Remote Healthbar via Find");
+                        return null;
+                    }
+                }
+                return healthbar;
+            }
+            else {
+                MelonLogger.Warning($"Unknown controller type: {controllerType}");
+                return null;
+            }
+        }
 
-        //        if (controller.controllerType == ControllerType.Local)
-        //        {
-        //            bool foundYet = false;
-        //            foreach (GameObject obj in Object.FindObjectsOfType<GameObject>()) // Removes clones
-        //            {
-        //                if (obj.name == "Player Controller(Clone)")
-        //                {
-        //                    if (foundYet)
-        //                        return;
-        //                    foundYet = true;
-        //                }
-        //            }
-        //        }
-
-        //        MelonLogger.Warning("Pass " + (controller.controllerType == ControllerType.Local));
-
-
-        //        if (controller.controllerType == ControllerType.Local)
-        //        {
-        //            MelonLogger.Warning("BLAH");
-        //        }
-        //    }
-        //}
+        [HarmonyPatch(typeof(PlayerHealth), nameof(PlayerHealth.Initialize))]
+        class PlayerHealthInitPatch
+        {
+            static void Postfix(PlayerHealth __instance, PlayerController controller)
+            {
+                AddHealthbarText(GetHealthbar(__instance.transform, controller.controllerType), controller.assignedPlayer);
+            }
+        }
 
         [HarmonyPatch(typeof(PlayerHealth), nameof(PlayerHealth.SetHealthBarPercentage))]
-        class HealthBarPercentagePatch
+        class SetHealthBarPercentagePatch
         {
-            static void Postfix(ref PlayerHealth __instance, float currentHealth)
+            static void Postfix(PlayerHealth __instance, float currentHealth)
             {
-                if (__instance.parentController?.controllerType == ControllerType.Local)
-                {
-                    OnLocalHealthChanged(currentHealth);
+                TextMeshPro textMaybeNull = GetHealthbar(__instance.transform, __instance.parentController?.controllerType)?.Find("HealthText")?.GetComponent<TextMeshPro>();
+                if (textMaybeNull != null) {
+                    textMaybeNull.text = currentHealth.ToString();
                 }
-            }
-        }
-
-        private static void OnLocalHealthChanged(float newHealth)
-        {
-            if (localHealthBarText != null)
-            {
-                localHealthBarText.text = newHealth.ToString();
-            }
-            else
-            {
-                MelonLogger.Error("Local UI Bar Text detected as null. Attempting to fix.");
-                MapInit();
             }
         }
     }