+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.1.3
+- No longer uses RMAPI.
+	- Gets everthing via HarmonyPatch.
+- Attempt to make health numbers show for remote players.
+	- Dosen't seem to work yet.
+
 # v0.1.2
 - Fix issues when the user is client.
 	- It now grabs the UI Bar from the local player instead of any player.